backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
torch
transformers
scikit-learn
numpy
scipy
nltk
python-dotenv
pydantic
